Added: Date to bar chart
